% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bash_merge.R
\name{bash_merge}
\alias{bash_merge}
\title{Merge all csv files within a directory into a single file}
\usage{
bash_merge(folder, awk = TRUE, joined_file = "0000-merged.csv")
}
\arguments{
\item{folder}{path to directory with csv files to be merged}

\item{awk}{use awk (default: TRUE) if there is a header row, else use cat}

\item{joined_file}{name of the merged file}
}
\value{
none
}
\description{
Merge all csv files within a directory into a single file
}
