% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grouping_cols.R
\name{create_grouping_cols}
\alias{create_grouping_cols}
\title{Create a grouping column based on specified time buffer}
\usage{
create_grouping_cols(df, buffer)
}
\arguments{
\item{df}{a dataframe with `utc_orig` or `UTC time` column}

\item{buffer}{a buffer (in seconds) between "groupings" of data collection}
}
\value{
a new dataframe with columns t_delta, time_gap, and grouping
}
\description{
Create a grouping column based on specified time buffer
}
