% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_dplyr_helpers.R
\name{create_metadata_cols}
\alias{create_metadata_cols}
\title{Takes a dataframe and file path and adds metadata to the dataframe}
\usage{
create_metadata_cols(df, file_path)
}
\arguments{
\item{df}{dataframe to add metadata to (study_name, user_id, data_type)}

\item{file_path}{path to the originating CSV file}
}
\value{
a dataframe with metadata columns
}
\description{
Takes a dataframe and file path and adds metadata to the dataframe
}
