% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_dplyr_helpers.R
\name{create_time_cols}
\alias{create_time_cols}
\title{Take in a dataframe with utc_orig and return date components as columns}
\usage{
create_time_cols(df)
}
\arguments{
\item{df}{dataframe with `utc_orig` column}
}
\value{
dataframe with year, month, day, hour, minute columns
}
\description{
Take in a dataframe with utc_orig and return date components as columns
}
