% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_dirs.R
\name{.list_dirs}
\alias{.list_dirs}
\title{A convenience wrapper for list.dirs() with our more common defaults}
\usage{
.list_dirs(folder, full_path = FALSE, recursive = FALSE)
}
\arguments{
\item{folder}{file path to the study folder}

\item{full_path}{return full paths (including study folder)? Default: FALSE}

\item{recursive}{check paths recursively? Default: FALSE}
}
\value{
a vector of user folders
}
\description{
A convenience wrapper for list.dirs() with our more common defaults
}
