% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_folders.R
\name{.list_user_folders}
\alias{.list_user_folders}
\title{Helper that returns all the user folders for any given study folder}
\usage{
.list_user_folders(study_folder, full_path = FALSE)
}
\arguments{
\item{study_folder}{file path to the study folder}

\item{full_path}{return full paths (including study folder)? Default: FALSE}
}
\value{
a vector of user folders
}
\description{
Helper that returns all the user folders for any given study folder
}
