% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_file_name.R
\name{.parse_file_name}
\alias{.parse_file_name}
\title{Given a file path (in the original format), returns metadata about path}
\usage{
.parse_file_name(csv_path, with_dttm = FALSE)
}
\arguments{
\item{csv_path}{path to the csv file as a string}

\item{with_dttm}{when TRUE, also returns the timestamp of the file}
}
\value{
a list
}
\description{
Beiwe data is structured as:
     /study_folder/user_folder/sensor_folder/DATE_TIME.csv
}
\details{
This is just a convenience function that extract important metadata from
the file path. Given any file path (as a string), it will return a list
with elements $data_type, $user_id, $study_name, $file_path, and optionally
$datetime.
}
