% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_time_only.R
\name{keep_time_only}
\alias{keep_time_only}
\title{Simply subsets any given dataframe to only columns with "time" in the name}
\usage{
keep_time_only(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a dataframe with only columns containing "time" in the name
}
\description{
Simply subsets any given dataframe to only columns with "time" in the name
}
