% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stripped_csv.R
\name{make_stripped_csv}
\alias{make_stripped_csv}
\title{Given a CSV file returns an identical file with only timestamp data}
\usage{
make_stripped_csv(csv_path, suffix = "_stripped")
}
\arguments{
\item{csv_path}{path to the csv file as a string}

\item{suffix}{string to append to new csv file}
}
\value{
none
}
\description{
In some instances, it may not be possible or desirable to merge all files
first. This function will allow a researcher to strip away data on a file-
specific basis. It is *much* slower than the functions that merge all files
before reading them in or saving them (e.g., bash_merge() or tidy_merge()).
}
