% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_time.R
\name{rename_time}
\alias{rename_time}
\title{Just rename the original time column (UTC time) to utc_orig}
\usage{
rename_time(df)
}
\arguments{
\item{df}{dataframe with `UTC time` column}
}
\value{
dataframe with utc_orig column
}
\description{
Wrapper function purely for readability of code in a pipeline-friendly
way.
}
