% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_csv.R
\name{summarize_csv}
\alias{summarize_csv}
\title{Return metadata about a csv file}
\usage{
summarize_csv(csv_path)
}
\arguments{
\item{csv_path}{path to the csv file as a string}
}
\value{
a tibble with metadata about that file
}
\description{
Given a file path, this function will return a dataframe with file metadata:
 study_name, user_id, data_type, timestamp, n_rows, and file_path
}
