% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_groupings.R
\name{summarize_groupings}
\alias{summarize_groupings}
\title{Summarize the groupings that you've already calculated}
\usage{
summarize_groupings(df)
}
\arguments{
\item{df}{a dataframe with grouping columns (see `create_grouping_cols()``)}
}
\value{
a summarized dataframe
}
\description{
Summarize the groupings that you've already calculated
}
