% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_merge.R
\name{tidy_merge}
\alias{tidy_merge}
\title{Combine csv files into a single dataframe, removing sensor data}
\usage{
tidy_merge(file_list, na = c("", "NA", "unknown"))
}
\arguments{
\item{file_list}{a vector of file paths to be combined}

\item{na}{a vector of values to read in as NA}
}
\value{
a tibble with timestamp and `UTC time` columns
}
\description{
This helper function will read in all csv files, remove sensor data, and
combine all files into a single dataframe.
}
